If merged, this commit will restructure our files to be more semantic and readable
